Use uint

+++ src/lib.rs
@@ -4,7 +4,7 @@
 
 extern crate pwasm_std;
 extern crate parity_hash as hash;
-extern crate bigint;
+extern crate uint;
 
 mod ext;
 mod storage;
